Move WildcardToRegex to Obfuscation namespace
